refactor(docs-pproduct): common files
